fix compile bug for merge mining & mpool
