--- pkg/chain/util.go
@@ -1,88 +1,0 @@
-package chain
-
-import (
-	"reflect"
-	"runtime"
-	"strings"
-
-	"github.com/filecoin-project/go-state-types/abi"
-	"github.com/filecoin-project/go-state-types/rt"
-	"github.com/filecoin-project/venus/pkg/block"
-	"github.com/ipfs/go-cid"
-
-	exported0 "github.com/filecoin-project/specs-actors/actors/builtin/exported"
-	exported2 "github.com/filecoin-project/specs-actors/v2/actors/builtin/exported"
-
-	"github.com/filecoin-project/venus/pkg/specactors/builtin"
-)
-
-type MethodMeta struct {
-	Name string
-
-	Params reflect.Type
-	Ret    reflect.Type
-}
-
-var MethodsMap = map[cid.Cid]map[abi.MethodNum]MethodMeta{}
-
-func init() {
-	// TODO: combine with the runtime actor registry.
-	var actors []rt.VMActor
-	actors = append(actors, exported0.BuiltinActors()...)
-	actors = append(actors, exported2.BuiltinActors()...)
-
-	for _, actor := range actors {
-		exports := actor.Exports()
-		methods := make(map[abi.MethodNum]MethodMeta, len(exports))
-
-		// Explicitly add send, it's special.
-		methods[builtin.MethodSend] = MethodMeta{
-			Name:   "Send",
-			Params: reflect.TypeOf(new(abi.EmptyValue)),
-			Ret:    reflect.TypeOf(new(abi.EmptyValue)),
-		}
-
-		// Iterate over exported methods. Some of these _may_ be nil and
-		// must be skipped.
-		for number, export := range exports {
-			if export == nil {
-				continue
-			}
-
-			ev := reflect.ValueOf(export)
-			et := ev.Type()
-
-			// Extract the method names using reflection. These
-			// method names always match the field names in the
-			// `builtin.Method*` structs (tested in the specs-actors
-			// tests).
-			fnName := runtime.FuncForPC(ev.Pointer()).Name()
-			fnName = strings.TrimSuffix(fnName[strings.LastIndexByte(fnName, '.')+1:], "-fm")
-
-			switch abi.MethodNum(number) {
-			case builtin.MethodSend:
-				panic("method 0 is reserved for Send")
-			case builtin.MethodConstructor:
-				if fnName != "Constructor" {
-					panic("method 1 is reserved for Constructor")
-				}
-			}
-
-			methods[abi.MethodNum(number)] = MethodMeta{
-				Name:   fnName,
-				Params: et.In(1),
-				Ret:    et.Out(0),
-			}
-		}
-		MethodsMap[actor.Code()] = methods
-	}
-}
-
-// Reverse reverses the order of the slice `chain`.
-func Reverse(chain []*block.TipSet) {
-	// https://github.com/golang/go/wiki/SliceTricks#reversing
-	for i := len(chain)/2 - 1; i >= 0; i-- {
-		opp := len(chain) - 1 - i
-		chain[i], chain[opp] = chain[opp], chain[i]
-	}
-}
