+++ pkg/consensus/call_invoke.go
@@ -5,25 +5,64 @@ import (
 	"fmt"
 
 	"github.com/filecoin-project/go-state-types/abi"
+	"github.com/ipfs/go-cid"
+	xerrors "github.com/pkg/errors"
+	"go.opencensus.io/trace"
+
 	"github.com/filecoin-project/venus/pkg/block"
 	"github.com/filecoin-project/venus/pkg/constants"
+	"github.com/filecoin-project/venus/pkg/fork"
 	"github.com/filecoin-project/venus/pkg/types"
 	"github.com/filecoin-project/venus/pkg/vm"
 	"github.com/filecoin-project/venus/pkg/vm/state"
-	xerrors "github.com/pkg/errors"
-	"go.opencensus.io/trace"
 )
 
-func (c *Expected) CallWithGas(ctx context.Context, msg *types.UnsignedMessage) (*vm.Ret, error) {
-	head := c.chainState.GetHead()
-	stateRoot, err := c.chainState.GetTipSetStateRoot(head)
+func (c *Expected) CallWithGas(ctx context.Context, msg *types.UnsignedMessage, priorMsgs []types.ChainMsg, ts *block.TipSet) (*vm.Ret, error) {
+	var (
+		err       error
+		stateRoot cid.Cid
+		height    abi.ChainEpoch
+	)
+	if ts == nil {
+		ts, err = c.chainState.GetTipSet(c.chainState.GetHead())
+		if err != nil {
+			return nil, err
+		}
+
+		// Search back till we find a height with no fork, or we reach the beginning.
+		// We need the _previous_ height to have no fork, because we'll
+		// run the fork logic in `sm.TipSetState`. We need the _current_
+		// height to have no fork, because we'll run it inside this
+		// function before executing the given message.
+		height, err = ts.Height()
+		if err != nil {
+			return nil, err
+		}
+		for height > 0 && (c.fork.HasExpensiveFork(ctx, height) || c.fork.HasExpensiveFork(ctx, height-1)) {
+			ts, err = c.chainState.GetTipSet(ts.EnsureParents())
+			if err != nil {
+				return nil, xerrors.Errorf("failed to find a non-forking epoch: %w", err)
+			}
+		}
+
+		stateRoot, err = c.chainState.GetTipSetStateRoot(c.chainState.GetHead())
+		if err != nil {
+			return nil, err
+		}
+	} else {
+		stateRoot, err = c.chainState.GetTipSetStateRoot(ts.Key())
+		if err != nil {
+			return nil, err
+		}
+	}
+
+	// When we're not at the genesis block, make sure we don't have an expensive migration.
+	height, err = ts.Height()
 	if err != nil {
 		return nil, err
 	}
-
-	ts, err := c.chainState.GetTipSet(head)
-	if err != nil {
-		return nil, err
+	if height > 0 && (c.fork.HasExpensiveFork(ctx, height) || c.fork.HasExpensiveFork(ctx, height-1)) {
+		return nil, fork.ErrExpensiveFork
 	}
 
 	vms := vm.NewStorage(c.bstore)
@@ -48,9 +87,17 @@ func (c *Expected) CallWithGas(ctx context.Context, msg *types.UnsignedMessage) 
 		NtwkVersionGetter: c.fork.GetNtwkVersion,
 		Rnd:               &rnd,
 		BaseFee:           ts.At(0).ParentBaseFee,
-		Epoch:             ts.At(0).Height,
+		Epoch:             height + 1,
 		GasPriceSchedule:  c.gasPirceSchedule,
 	}
+
+	for i, m := range priorMsgs {
+		_, err := c.processor.ProcessUnsignedMessage(ctx, m.VMMessage(), priorState, vms, vmOption)
+		if err != nil {
+			return nil, xerrors.Errorf("applying prior message (%d): %w", i, err)
+		}
+	}
+
 	return c.processor.ProcessUnsignedMessage(ctx, msg, priorState, vms, vmOption)
 }
 
@@ -58,7 +105,6 @@ func (c *Expected) Call(ctx context.Context, msg *types.UnsignedMessage, ts *blo
 	ctx, span := trace.StartSpan(ctx, "statemanager.Call")
 	defer span.End()
 	chainReader := c.chainState
-	fork := c.fork
 	// If no tipset is provided, try to find one without a fork.
 	var err error
 	if ts == nil {
@@ -69,7 +115,7 @@ func (c *Expected) Call(ctx context.Context, msg *types.UnsignedMessage, ts *blo
 		}
 
 		// Search back till we find a height with no fork, or we reach the beginning.
-		for ts.EnsureHeight() > 0 && fork.HasExpensiveFork(ctx, ts.EnsureHeight()-1) {
+		for ts.EnsureHeight() > 0 && c.fork.HasExpensiveFork(ctx, ts.EnsureHeight()-1) {
 			var err error
 			ts, err = chainReader.GetTipSet(ts.EnsureParents())
 			if err != nil {
@@ -89,12 +135,12 @@ func (c *Expected) Call(ctx context.Context, msg *types.UnsignedMessage, ts *blo
 	// return an error because the migration will take too long.
 	//
 	// We allow this at height 0 for at-genesis migrations (for testing).
-	if bheight-1 > 0 && fork.HasExpensiveFork(ctx, bheight-1) {
+	if bheight-1 > 0 && c.fork.HasExpensiveFork(ctx, bheight-1) {
 		return nil, ErrExpensiveFork
 	}
 
 	// Run the (not expensive) migration.
-	bstate, err = fork.HandleStateForks(ctx, bstate, bheight-1, ts)
+	bstate, err = c.fork.HandleStateForks(ctx, bstate, bheight-1, ts)
 	if err != nil {
 		return nil, fmt.Errorf("failed to handle fork: %w", err)
 	}
@@ -128,7 +174,7 @@ func (c *Expected) Call(ctx context.Context, msg *types.UnsignedMessage, ts *blo
 			}
 			return dertail.FilCirculating, nil
 		},
-		NtwkVersionGetter: fork.GetNtwkVersion,
+		NtwkVersionGetter: c.fork.GetNtwkVersion,
 		Rnd:               &rnd,
 		BaseFee:           ts.At(0).ParentBaseFee,
 		Epoch:             ts.At(0).Height,
